requirements.txt
flask
pymongo
pydantic
python-dotenv

